requirements.txt
python-dateutil==2.9.0
beautifulsoup4==4.12.3
requests==2.31.0